backend/requirements.txt
Flask
groq
python-dotenv
flask-cors
google-generativeai
web3



